refactor: rename text-animate to pascal case
