+++ app/page.tsx
@@ -9,7 +9,7 @@ import { getEventData } from "@/app/queries"
 import { getStringDate } from "@/components/calendar/utils"
 import React, { Suspense } from "react"
 import Spinner from "@/components/assets/Spinner"
-import { TextAnimate } from "@/components/magicui/text-animate"
+import { TextAnimate } from "@/components/magicui/TextAnimate"
 import { Button } from "@/components/ui/Button"
 import { ScrollButton } from "@/components/ui/ScrollButton"
